+++ packages/mcp/src/utils.ts
@@ -1,0 +1,15 @@
+import { ServiceError } from "./types.js";
+
+// From https://developer.mozilla.org/en-US/docs/Glossary/Base64#the_unicode_problem
+export const base64Decode = (base64: string): string => {
+    const binString = atob(base64);
+    return Buffer.from(Uint8Array.from(binString, (m) => m.codePointAt(0)!).buffer).toString();
+}
+
+export const isServiceError = (data: unknown): data is ServiceError => {
+    return typeof data === 'object' &&
+        data !== null &&
+        'statusCode' in data &&
+        'errorCode' in data &&
+        'message' in data;
+}